# api/openapi/atlas.openapi.json
{
  "components": {
    "schemas": {
...
      "EntityCollectionResponse": {
        "description": "Collection response for entities.",
        "properties": {
          "facets": {
            "anyOf": [
              {
                "additionalProperties": {
                  "items": {
                    "$ref": "#/components/schemas/FacetOption"
                  },
                  "type": "array"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Facets"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/EntityResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "place": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Address"
              },
              {
                "type": "null"
              }
            ]
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total"
        ],
        "title": "EntityCollectionResponse",
        "type": "object"
      },
...
      "EntityDetailResponse": {
        "description": "Expanded entity detail record.",
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "address": {
            "$ref": "#/components/schemas/Address"
          },
          "affiliated_org_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Affiliated Org Id"
          },
          "contact": {
            "$ref": "#/components/schemas/ContactInfo"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "flag_summary": {
            "$ref": "#/components/schemas/FlagSummary"
          },
          "freshness": {
            "$ref": "#/components/schemas/FreshnessInfo"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "issue_area_ids": {
            "items": {
              "type": "string"
            },
            "title": "Issue Area Ids",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "relationship_ids": {
            "items": {
              "type": "string"
            },
            "title": "Relationship Ids",
            "type": "array"
          },
          "resource_uri": {
            "title": "Resource Uri",
            "type": "string"
          },
          "source_count": {
            "default": 0,
            "title": "Source Count",
            "type": "integer"
          },
          "source_ids": {
            "items": {
              "type": "string"
            },
            "title": "Source Ids",
            "type": "array"
          },
          "source_types": {
            "items": {
              "type": "string"
            },
            "title": "Source Types",
            "type": "array"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceResponse"
            },
            "title": "Sources",
            "type": "array"
          },
          "type": {
            "title": "Type",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          },
          "verified": {
            "title": "Verified",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "type",
          "name",
          "description",
          "address",
          "contact",
          "active",
          "verified",
          "freshness",
          "created_at",
          "updated_at",
          "resource_uri"
        ],
        "title": "EntityDetailResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "contact": {
      "email": "atlas@example.com",
      "name": "Rebuilding America",
      "url": "https://atlas.example.com/"
    },
    "description": "Atlas exposes place-first civic research data through a typed REST API. The public contract includes entities, sources, issue areas, domains, place profiles, issue signals, discovery runs, and anonymous data flags.",
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "summary": "Structured civic research data for Atlas places, entities, sources, and discovery runs.",
    "title": "Atlas REST API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/entities": {
      "get": {
        "description": "Search Atlas entities with text, geography, issue-area, entity-type, and source-type filters.",
        "operationId": "listEntities",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Query"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "in": "query",
            "name": "city",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "City"
            }
          },
          {
            "in": "query",
            "name": "region",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Region"
            }
          },
          {
            "in": "query",
            "name": "entity_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Type"
            }
          },
          {
            "in": "query",
            "name": "issue_area",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Issue Area"
            }
          },
          {
            "in": "query",
            "name": "source_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Type"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityCollectionResponse"
                }
              }
            },
            "description": "A paginated collection of Atlas entities."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List entities",
        "tags": [
          "entities"
        ]
      },
      "post": {
        "description": "Create a new Atlas entity using the canonical nested address and contact request shape.",
        "operationId": "createEntity",
        "parameters": [
          {
            "in": "header",
            "name": "x-atlas-internal-secret",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Atlas-Internal-Secret"
            }
          },
          {
            "in": "header",
            "name": "x-atlas-actor-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Atlas-Actor-Id"
            }
          },
          {
            "in": "header",
            "name": "x-atlas-actor-email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Atlas-Actor-Email"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntityCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityDetailResponse"
                }
              }
            },
            "description": "The newly created Atlas entity."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create an entity",
        "tags": [
          "entities"
        ]
      }
    },
    "/api/entities/{entity_id}": {
      "delete": {
        "description": "Delete an Atlas entity by ID.",
        "operationId": "deleteEntity",
        "parameters": [
          {
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "title": "Entity Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-atlas-internal-secret",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Atlas-Internal-Secret"
            }
          },
          {
            "in": "header",
            "name": "x-atlas-actor-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Atlas-Actor-Id"
            }
          },
          {
            "in": "header",
            "name": "x-atlas-actor-email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Atlas-Actor-Email"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The entity was deleted."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete an entity",
        "tags": [
          "entities"
        ]
      },
      "get": {
        "description": "Return one Atlas entity with normalized contact, freshness, flag summary, and linked source provenance.",
        "operationId": "getEntity",
        "parameters": [
          {
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "title": "Entity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityDetailResponse"
                }
              }
            },
            "description": "The requested Atlas entity."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get an entity",
        "tags": [
          "entities"
        ]
      },
      "patch": {
        "description": "Apply a partial update to an Atlas entity.",
        "operationId": "updateEntity",
        "parameters": [
          {
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "title": "Entity Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-atlas-internal-secret",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Atlas-Internal-Secret"
            }
          },
          {
            "in": "header",
            "name": "x-atlas-actor-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Atlas-Actor-Id"
            }
          },
          {
            "in": "header",
            "name": "x-atlas-actor-email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Atlas-Actor-Email"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntityUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityDetailResponse"
                }
              }
            },
            "description": "The updated Atlas entity."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update an entity",
        "tags": [
          "entities"
        ]
      }
    },
    "/api/entities/{entity_id}/sources": {
      "get": {
        "description": "Return the source trail for one Atlas entity.",
        "operationId": "listEntitySources",
        "parameters": [
          {
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "title": "Entity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntitySourcesResponse"
                }
              }
            },
            "description": "The source records linked to the requested entity."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List entity sources",
        "tags": [
          "entities"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "https://atlas.example.com"
    },
    {
      "url": "http://localhost:8000"
    },
    {
      "url": "/"
    }
  ],
...
}